refactor: 카카오 로그인 로직 AuthFacadeService로 분리
- 기존 로직은 트랜잭션 범위 밖에서 지연로딩된 엔티티에 접근해 LazyInitializationException 가 발생했음.
- 클래스를 분리해서 DB 작업과 외부 API 호출을 분리하였고, DB 작업은 트랜잭션을 보장해주도록 변경
- 파사드 클래스를 추가해 흐름을 제어할 수 있도록 함.
